Release (1)

+++ trabalho1 - SO/main.c
@@ -1,6 +1,6 @@
 // https://www-di.inf.puc-rio.br/~endler/courses/inf1316/
 /*
-Alunos: Luiz Eduardo Raffaini (2220982) & Isabela Braconi ()
+Alunos: Luiz Eduardo Raffaini (2220982) & Isabela Braconi (2312065)
 Professor: Endler 
 */
 
@@ -13,6 +13,8 @@ Professor: Endler
 #include <string.h>
 #include <sys/shm.h>
 
+int* terminados;
+
 void controller() {
     int pidKernel = getppid(), auxRand;
     srand(time(NULL) ^ (getpid() << 16));
@@ -20,7 +22,7 @@ void controller() {
     for(;;) {
         usleep(500000); // 0.5 segundos
         kill(pidKernel, SIGHUP); // envia iqr0
-        auxRand = rand() % 10000;
+        auxRand = rand() % 100;
         if (auxRand <= 10 && auxRand > 5) {
             kill(pidKernel, SIGUSR1); // envia iqr1
         } else if (auxRand <= 5) {
@@ -28,8 +30,6 @@ void controller() {
         }
     }
 }
-
-#define MAX 100
 
 void processo(int pidKernel, int memoriaCompartilhadaD, int memoriaCompartilhadaF) {
     raise(SIGSTOP);
@@ -45,12 +45,20 @@ void processo(int pidKernel, int memoriaCompartilhadaD, int memoriaCompartilhada
 
     raise(SIGSTOP);
 
-    int PC = 0;
-    while (PC < 100) {
+    int PC = 0, qtdD1 = 0, qtdD2 = 0;
+    while (PC < 10) {
         int random = rand() % 100;
         char dispotivo, func;
         if (random < 15) {
-            dispotivo = (random % 2) ? '1' : '2';
+            dispotivo = (random % 2);
+            if(dispotivo == 1){
+                dispotivo = '1';
+                qtdD1++;
+            }else{
+                dispotivo = '2';
+                qtdD2++;
+            }
+
             if (random % 2) {
                 func = 'r';
             } else if (random % 3) {
@@ -60,14 +68,16 @@ void processo(int pidKernel, int memoriaCompartilhadaD, int memoriaCompartilhada
             }
             *accessoMemoriaD = dispotivo; 
             *accessoMemoriaF = func;
-            printf("Dispositivo: %c : Func %c : sysCallProcess %c%c\n", dispotivo, func, *accessoMemoriaD, *accessoMemoriaF);
             kill(pidKernel, SIGINT);
             raise(SIGSTOP);
         }
         PC++;
     }
-    printf("processo terminado!\n");
     kill(pidKernel, SIGTERM);
+    printf("Processo %d usou %d o dispositivio D1 e %d o dispositivio D2\n", (getpid()),  qtdD1, qtdD2);
+    printf("processo terminado!\n");
+    
+    terminados[pidKernel - getpid() - 1] = 0;
 
     // Desanexar a memória compartilhada
     if (shmdt(accessoMemoriaD) == -1 || shmdt(accessoMemoriaF) == -1) {
@@ -92,10 +102,10 @@ void kernelHandler(int signal) {
     } else if (signal == SIGHUP) {
         irq0 = 1;
     } else if (signal == SIGUSR1) {
-       // printf("Kernel recebeu um sinal IRQ1 do controller...\n");
+        printf("Kernel recebeu um sinal I/O do D1 do controller...\n");
         irq1 = 1;
     } else if (signal == SIGUSR2) {
-        //printf("Kernel recebeu um sinal IRQ2 do controller...\n");
+        printf("Kernel recebeu um sinal I/O do D2 do controller...\n");
         irq2 = 1;
     }
 }
@@ -196,8 +206,8 @@ int main(void) {
     signal(SIGUSR1, kernelHandler);
     signal(SIGUSR2, kernelHandler);
 
-    int memoriaDispositivo = shmget(IPC_PRIVATE, sizeof(char), IPC_CREAT | 0666);
-    int memoriaFuncao = shmget(IPC_PRIVATE, sizeof(char), IPC_CREAT | 0666);
+    int memoriaDispositivo = shmget(IPC_PRIVATE, sizeof(char), IPC_CREAT | S_IRWXU);
+    int memoriaFuncao = shmget(IPC_PRIVATE, sizeof(char), IPC_CREAT | S_IRWXU);
     if (memoriaDispositivo == -1 || memoriaFuncao == -1) {
         perror("Erro ao criar memória compartilhada");
         exit(1);
@@ -216,6 +226,12 @@ int main(void) {
             processo(pid, memoriaDispositivo, memoriaFuncao);
             exit(0);
         }
+    }
+
+    terminados = (int*)malloc(sizeof(int) * nProcessos);
+    for (int i = 0; i < nProcessos; i++)
+    {
+        terminados[i] = 1;
     }
 
     int pidController = fork();
@@ -246,38 +262,45 @@ int main(void) {
             printf("Syscall: %c%c\n", *accessoMemoriaD, *accessoMemoriaF);
             estados[processoAtual] = 1;
             enFila(fila, *accessoMemoriaD - '0', processoAtual);
-            irq0 = 1;
             sysCallProcessFlag = 0; // Reseta a flag de sysCallProcess após processar
         }
+
         if (irq1) {
             Node* node = peek(fila);
-            if (node->dispositivo == 1) {
+            if (node != NULL && node->dispositivo == 1) {
                 estados[node->processo] = 0;
                 kill(processos[node->processo], SIGCONT);
                 free(deFila(fila));
+                printf("Liberando fila...\n");
             }
             irq1 = 0;
         }
+
         if (irq2) {
             Node* node = peek(fila);
-            if (node->dispositivo == 2) {
+            if (node != NULL && node->dispositivo == 2) {
                 estados[node->processo] = 0;
                 kill(processos[node->processo], SIGCONT);
                 free(deFila(fila));
+                printf("Liberando fila...\n");
             }
             irq2 = 0;
         }
+        
         if (irq0) {
             kill(processos[processoAtual], SIGSTOP);
+
+            processoAtual = (processoAtual + 1) % nProcessos;
             for (int i = 0; i < 3; i++)
             {
                 if (estados[processoAtual] == 0) {
+                    if (terminados[processoAtual]){
+                        kill(processos[processoAtual], SIGCONT);
                     break;
+                    }
                 }
-                processoAtual = (processoAtual + 1) % nProcessos;
             }
-            
-            kill(processos[processoAtual], SIGCONT);
+
             irq0 = 0;
         }
         if (processEndFlag == nProcessos) {
@@ -285,8 +308,6 @@ int main(void) {
             kill(pidController, SIGKILL);
             break;
         }
-        // Outras operações do kernel
-        sleep(1);
     }
 
     // Desanexar a memória compartilhada
